+++ src/types/CatalogProduct.ts
@@ -20,7 +20,7 @@ export interface CatalogProduct {
   resolution: string;
   processor: string;
   ram: string;
-  camera: string;
-  zoom: string;
+  camera?: string;
+  zoom?: string;
   cell: string[];
 }
